backend/requirements.txt
fastapi==0.116.1
httpx==0.28.1
PyYAML==6.0.3
uvicorn==0.35.0
